First Commit
This the BETA version of the ARabic Desktop Toolbar. The toolbar
provides accesibility features to arab desktop users. This is a visual
studio 2010 solution with four projects. The accessibilitytoolbar
project is the startup project. The colourchooserCS is a simple colour
wheel control. The two other projects are c# wrappers around windows api
for magnification and hooking to keyboard and mouse events